backend/requirements.txt
fastapi==0.110.0
hypercorn==0.16.0
tortoise-orm==0.20.0
asyncpg==0.29.0
pydantic==2.6.1
werkzeug==3.0.1
python-dotenv==1.0.0
watchdog==6.0.0
pydantic[email]==2.6.1
websockets==10.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
